Switch to env only settings
As of 4.1 cantaloupe supports changing all settings through environment
variables. This removes the templating solution we had in place as it's
no longer needed. The properties file that is added to the docker image
is the stock, unmodified version that is distributed with cantaloupe.

This commit also adds a docker-compose file for local testing of
changes. The compose configuration uses minio and more closely tracks
our production deployment to enable better testing.